requirements.txt
deepl
python-dotenv